refactor(monolith): extract e2e/conftest.py into shared/testing/plugin.py
Move all shared fixtures (PostgreSQL, uvicorn, SvelteKit, Playwright) to
shared.testing.plugin as the canonical location. The e2e/conftest.py becomes
a thin shim that delegates via pytest_plugins. Also registers BDD coverage
markers (covers_route, covers_page, covers_public) in pytest_configure.
